Render watchlist items on search screen
